Add SpinitronMetadata
This commit adds the SpinitronMetadata type, which represents the body
of the metadata push data sent from spinitron to our api. Note that the
contents of the metadata push are entirely customizable, documentation
can be found here
(https://forum.spinitron.com/t/metadata-push-guide/144).  Since this
type will be used by both the WNYU front-end and back-end, it is being
commited to the repository